Move the profiler to Common
