+++ Core/CoreTiming.cpp
@@ -20,7 +20,7 @@
 #include <cstdio>
 #include <mutex>
 
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Common/Serialize/Serializer.h"
 #include "Common/Serialize/SerializeList.h"
