+++ Core/HLE/HLE.cpp
@@ -20,7 +20,7 @@
 #include <vector>
 #include <string>
 
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Common/Log.h"
 #include "Common/Serialize/SerializeFuncs.h"
