+++ Core/HLE/sceIo.cpp
@@ -20,7 +20,7 @@
 #include <thread>
 
 #include "Common/Thread/ThreadUtil.h"
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Common/FileUtil.h"
 #include "Common/Serialize/SerializeFuncs.h"
