+++ Core/HW/SasAudio.cpp
@@ -17,7 +17,7 @@
 
 #include <algorithm>
 
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Common/Serialize/SerializeFuncs.h"
 #include "Core/MemMapHelpers.h"
