+++ Core/MIPS/x86/Jit.cpp
@@ -22,7 +22,7 @@
 #include <iterator>
 
 #include "Common/Math/math_util.h"
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Common/Serialize/Serializer.h"
 #include "Common/Serialize/SerializeFuncs.h"
