+++ GPU/Common/TextureCacheCommon.cpp
@@ -18,7 +18,7 @@
 #include <algorithm>
 
 #include "ppsspp_config.h"
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 #include "Common/ColorConv.h"
 #include "Common/MemoryUtil.h"
 #include "Core/Config.h"
