+++ GPU/D3D11/GPU_D3D11.cpp
@@ -40,7 +40,7 @@
 #include "Common/Serialize/Serializer.h"
 #include "Common/GraphicsContext.h"
 #include "base/NativeApp.h"
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 #include "Common/Data/Text/I18n.h"
 #include "Core/Debugger/Breakpoints.h"
 #include "Core/MemMapHelpers.h"
