+++ GPU/GLES/DrawEngineGLES.cpp
@@ -26,7 +26,7 @@
 #include "Core/CoreTiming.h"
 
 #include "gfx/gl_debug_log.h"
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "GPU/Math3D.h"
 #include "GPU/GPUState.h"
