+++ GPU/GLES/TextureCacheGLES.cpp
@@ -22,7 +22,7 @@
 #include "gfx/gl_debug_log.h"
 #include "Common/Data/Text/I18n.h"
 #include "Common/Math/math_util.h"
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 #include "thin3d/GLRenderManager.h"
 
 #include "Common/ColorConv.h"
