+++ GPU/Software/Clipper.cpp
@@ -23,7 +23,7 @@
 #include "GPU/Software/Rasterizer.h"
 #include "GPU/Software/RasterizerRectangle.h"
 
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 namespace Clipper {
