+++ GPU/Software/Rasterizer.cpp
@@ -18,7 +18,7 @@
 #include <algorithm>
 #include <cmath>
 
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Common/ThreadPools.h"
 #include "Common/ColorConv.h"
