+++ GPU/Software/RasterizerRectangle.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <cmath>
 
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Core/System.h"
