+++ GPU/Vulkan/GPU_Vulkan.cpp
@@ -18,7 +18,7 @@
 
 #include <thread>
 
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Common/Log.h"
 #include "Common/FileUtil.h"
