+++ GPU/Vulkan/PipelineManagerVulkan.cpp
@@ -3,7 +3,7 @@
 #include <set>
 #include <sstream>
 
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 #include "Common/Log.h"
 #include "Common/StringUtils.h"
