+++ GPU/Vulkan/TextureCacheVulkan.cpp
@@ -22,7 +22,7 @@
 #include "Common/File/VFS/VFS.h"
 #include "Common/Data/Text/I18n.h"
 #include "Common/Math/math_util.h"
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 #include "thin3d/thin3d.h"
 #include "thin3d/VulkanRenderManager.h"
