+++ UI/ProfilerDraw.cpp
@@ -21,7 +21,7 @@
 #include "gfx_es2/draw_buffer.h"
 #include "ui/ui_context.h"
 #include "ui/view.h"
-#include "profiler/profiler.h"
+#include "Common/Profiler/Profiler.h"
 
 static const uint32_t nice_colors[] = {
 	0xFF8040,
